requirements.txt
discord~=2.2.3
pymongo~=4.3.3
python-dotenv~=1.0.0
